+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Fix `IndexOutOfRangeException` when parsing SQL statements.
   ([#4139](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4139))
 
+* Add instrumentation schema URL to traces.
+  ([#TODO](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/TODO))
+
 ## 1.15.0-beta.1
 
 Released 2026-Jan-21